Revert "Merge remote-tracking branch 'upstream/dev' into xenos_acid"
This reverts commit a737caafd6fb0b39d6369c844185a488dde1b7da, reversing
changes made to 5aa8536612003d807617cadf096913cf76c2a9ef.

+++ code/game/objects/items/storage/specialized.dm
@@ -103,13 +103,9 @@
 			inserted = 1
 			break
 
-	if(!inserted || !S.amount)
+	if((!inserted || !S.amount) && usr.drop(S, src))
 		if(!S.amount)
 			qdel(S)
-		else if(S.loc == usr)
-			usr.drop(S, src)
-		else
-			S.forceMove(src)
 		usr.update_icons()	//update our overlays
 
 	prepare_ui(usr)
